Replace functional tests with molecule
As effort to sunset tests repo, we replace functional test that were
running for the role with a molecule.
Due to docker limitations, I had to comment out swap tests, as while
it's possible to create a container with swap, I didn't find a way to
`swapon` from inside of the container.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/938571
Change-Id: I80297a01ff6e820591325529c3457302e021469a