Replace block/rescue with ignore_errors/register
Users are always getting triggered by the failed
`systemctl reload-or-restart` command. To potentially reduce confusion,
let's replace block/rescue back to `ignore_errors`, as in this case
Ansible at least printing out `...ignoring` for the raised error,
instead of silently proceeding further, which might help out
to see that the failure is non-critical.

Change-Id: I264e23598dd211c5ddef971d2cc3702f4f2af3b6